Factor test scripts out of groovy file, whitelist users
Negative test for disallowed user:
https://fdb-jenkins.int.snowflakecomputing.com/view/Github/job/snowflakedb/job/foundationdb/view/change-requests/job/PR-30/11/console
    + set +x
    User marcmac is NOT on the whitelist - Aborting

    GitHub has been notified of this commit’s build result

    Finished: ABORTED

Positive test for allowed user:
https://fdb-jenkins.int.snowflakecomputing.com/view/Github/job/snowflakedb/job/foundationdb/view/change-requests/job/PR-30/12/console
    "User marcmac is on the whitelist"

Tests that are refused will be marked as aborted, not failed.

Note that the testing above was done with the PR config relaxed, to allow the new logic to be tested.
In the future, changes to this will have to be added to the jenkinsfile via a PR.

SNOW-80477
@review pr in github